Updated code to append cons.csv, refactored for readability

+++ Infer/handle_data.py
@@ -8,7 +8,7 @@ import json
 class HandleData():
     def __init__(self, aws_id, aws_secret, bucket, startfromscratch, custids, startdate, mode='r'):
         self.client = boto3.client('s3', aws_access_key_id=aws_id,
-                              aws_secret_access_key=aws_secret)
+                                   aws_secret_access_key=aws_secret)
         self.mode = mode
         self.startfromscratch = startfromscratch
         self.bucket = bucket
@@ -17,7 +17,8 @@ class HandleData():
             self.dct_inp['state'] = {}
             self.dct_inp['state']['last_run_date'] = pd.to_datetime(startdate, format='%d/%m/%Y')
             self.dct_inp['state']['cust_notified_already'] = set()
-            self.dct_inp['state']['Cumulative_Consumption'] = pd.DataFrame(0, index=custids, columns=['Cumulative_Consumption']).to_json()
+            self.dct_inp['state']['Cumulative_Consumption'] = pd.DataFrame(0, index=custids,
+                                                                           columns=['Cumulative_Consumption']).to_json()
 
     def __read_json(self, bucket_name, object_key):
         content_object = self.client.get_object(bucket_name, object_key)
@@ -34,7 +35,7 @@ class HandleData():
 
     def __write_json(self, bucket_name, object_key, file):
         serializedMyData = json.dumps(file)
-        self.client.put_object(Bucket=bucket_name, Key=object_key, Body = serializedMyData)
+        self.client.put_object(Bucket=bucket_name, Key=object_key, Body=serializedMyData)
 
     def __write_csv(self, bucket_name, object_key, file):
         csv_buffer = StringIO()
@@ -57,7 +58,7 @@ class HandleData():
     def read_write_main(self, state=None, cons=None):
         if self.mode == 'r':
             self.dct_inp['coeff'] = self._handle_file_rw(self.bucket, coeff_key, 'csv')
-            if not(self.startfromscratch):
+            if not (self.startfromscratch):
                 self.dct_inp['state'] = self._handle_file_rw(self.bucket, state_key, 'json')
             self.dct_inp['order'] = self._handle_file_rw(self.bucket, order_key, 'csv')
             self.dct_inp['cust'] = self._handle_file_rw(self.bucket, cust_key, 'csv')
@@ -67,14 +68,13 @@ class HandleData():
             self._handle_file_rw(self.bucket, state_key, 'json', state)
             self._handle_file_rw(self.bucket, cons_key, 'csv', cons)
 
-
     def read_write_main_local(self, state=None, cons=None):
         if self.mode == 'r':
-            if not(self.startfromscratch):
+            if not (self.startfromscratch):
                 self.dct_inp = {}
-                with open(r'./Data/State/notify_simp_avg.json', 'r') as fr:
+                with open(r'./Data/State/notify_2', 'r') as fr:
                     self.dct_inp['state'] = json.load(fr)
-            self.dct_inp['coeff'] = pd.read_csv(r'./Data/Input/Cust_coeff_simple_avg.csv')
+            self.dct_inp['coeff'] = pd.read_csv(r'./Data/Input/Cust_coeff2.csv')
             self.dct_inp['order'] = pd.read_csv(r'./Data/Input/OrderData1.csv')
             try:
                 self.dct_inp['cust'] = pd.read_csv(r'./Data/Input/Customer_First_Entry_Date.csv', header=None)
@@ -86,6 +86,30 @@ class HandleData():
             self.dct_inp['temp'] = pd.read_csv(r'./Data/Input/temperature1.csv')
             return self.dct_inp
         elif self.mode == 'w':
-            with open(r'./Data/State/notify_simp_avg.json', 'w') as fw:
+            with open(r'./Data/State/notify_2.json', 'w') as fw:
                 json.dump(state, fw, indent=4)
-            cons.to_csv(r'./Data/Debug/cons.csv')
+            cons.index.name = 'Customers'
+            cons.columns.name = 'Date'
+            cons = cons.T
+            if os.path.isfile(r'./Data/Debug/cons.csv'):
+                df_cons = pd.read_csv(r'./Data/Debug/cons.csv')
+                df_cons['Date'] = pd.to_datetime(df_cons['Date'], format='%Y-%m-%d')
+                df_cons.set_index('Date', inplace=True)
+                df_cons.columns = [int(col) for col in df_cons.columns]
+                df_cons = df_cons.append(cons)
+                df_cons = df_cons[~df_cons.index.duplicated(keep='first')]
+                df_cons.sort_index(inplace=True)
+                df_cons.to_csv(r'./Data/Debug/cons.csv')
+            else:
+                cons.to_csv(r'./Data/Debug/cons.csv')
+            #     try:
+            #         df_cons = df_cons.set_index('Customers').T
+            #     except KeyError:
+            #         df_cons = df_cons.T.set_index('Customers')
+            #     cons.index.name = 'Customers'
+            #     df_cons = df_cons.append(cons.T)
+            #     df_cons.dropna(how='all', inplace=True)
+            #     df_cons.to_csv(r'./Data/Debug/cons.csv')
+            # except FileNotFoundError:
+            #     cons.index.name = 'Customers'
+            #     cons.to_csv(r'./Data/Debug/cons.csv')
